added git submodule for Scinterm
this is actually referencing the branch `sciteco-dev` of the
Scinterm Git mirror at https://github.com/rhaberkorn/scinterm-mirror.git
This currently corresponds to Scinterm v1.6 with some
not yet released upstream commits and custom patches that are not
yet upstream.

When preparing a SciTECO release, separate tags per SciTECO
release are created in the scintilla-mirror and scinterm-mirror
repositories that ease following which SciTECO version used which
state of Scintilla/Scinterm and ensure that commits are not discarded
in the mirror repositories even if their histories (the `sciteco-dev`
branches) are rewritten.